# src/main/resources/static/api-docs.yml
openapi: 3.0.0
info:
  title: "authorization-registry-service"
  version: 1.0.0
  description: >-
    Business management service.
  license:
    name: Apache-2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
paths:
...
  /restaurant/{restaurant_id}:
    get:
      tags:
        - restaurant
      description: Find a restaurant
      operationId: findById
      parameters:
        - name: restaurant_id
          in: path
          description: Id of a specific restaurant
          required: true
          schema:
            type: string
      responses:
        200:
          description: Restaurant created successfully
          content:
            application/vnd.anon_restaurant_v1+json:
              schema:
                $ref: '#/components/schemas/RestaurantResponseV1'
...
components:
  schemas:
...
    RestaurantResponseV1:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        delivery_fee:
          type: BigDecimal
        items:
          type: array
          items:
            $ref: '#/components/schemas/ItemResponseV1'
      required:
        - name
        - email
        - phone
        - address
        - delivery_fee
        - items
    ItemResponseV1:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: BigDecimal
        category:
          $ref: '#/components/schemas/CategoryV1'
      required:
        - name
        - description
        - price
    CategoryV1:
      type: object
      properties:
        name:
          type: string
      required:
        - name
...
  securitySchemes:
    CustomerAuth:
      type: http
      scheme: bearer
    ServiceAuth:
      type: apiKey
      name: Authorization
      in: header